+++ test/kelvin_waves.jl
@@ -98,7 +98,7 @@ function test_kelvin_waves(
             backend_short = NaiveShortRangeBackend(),
             backend_long = NonuniformFFTsBackend(m = HalfSupport(5), σ = 1.5),  # => 1e-8 accuracy
             quadrature = quad,
-            lia_segment_fraction = 0.1,
+            lia_segment_fraction = quad === NoQuadrature() ? nothing : 0.1,
         )
     end